Agrego metodos crud al service de personas

+++ src/main/java/com/gestionpersonas/GestionPersonas/repository/PersonaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonaRepository extends CrudRepository<Persona, Integer> {
+public interface PersonaRepository extends CrudRepository<Persona, Long> {
 }
